Move from `content-twitter/lib/tweet.ts` to `lib/tweet/tweet.ts`
Remove old interface Tweet.
