+++ src/lib/tweet/tweet.ts
@@ -1,7 +1,6 @@
 export type TweetID = string;
 
 export interface User {
-  id: string;
   name: string;
   username: string;
 }
@@ -14,25 +13,17 @@ export interface TweetEntityText {
 export interface TweetEntityURL {
   type: 'url';
   text: string;
-  url: string;
-  display_url: string;
+  short_url: string;
+  expanded_url: string;
   decoded_url: string;
   title?: string;
-  description?: string;
-}
-
-export interface TweetEntityMedia {
-  type: 'media';
-  text: string;
-  media_key: string;
-  media_type: 'animated_gif' | 'photo' | 'video' | string;
-  url?: string;
 }
 
 export interface TweetEntityHashtag {
   type: 'hashtag';
   text: string;
   tag: string;
+  hashmoji?: string;
 }
 
 export interface TweetEntityCashtag {
@@ -44,39 +35,54 @@ export interface TweetEntityCashtag {
 export interface TweetEntityMention {
   type: 'mention';
   text: string;
-  user: {
-    id: string;
-    name?: string;
-    username: string;
-  };
-}
-
-export interface TweetEntityAnnotation {
-  type: 'annotation';
-  text: string;
-  probability: number;
-  annotation_type: string;
-  normalized_text: string;
+  username: string;
 }
 
 export type TweetEntity =
   | TweetEntityText
   | TweetEntityURL
-  | TweetEntityMedia
   | TweetEntityHashtag
   | TweetEntityCashtag
   | TweetEntityMention;
 
-export interface ReferencedTweet {
-  type: 'retweeted' | 'quoted' | 'replied_to';
-  id: TweetID;
+export interface MediaPhoto {
+  type: 'photo';
+  url: string;
 }
+
+export interface MediaVideo {
+  type: 'video';
+  thumbnail: string;
+}
+
+export type Media = MediaPhoto | MediaVideo;
+
+export interface CardLink {
+  url: string;
+  expanded_url: string;
+  decoded_url: string;
+  title?: string;
+}
+
+export interface CardSingle {
+  type: 'single';
+  link?: CardLink;
+  media_url: string;
+}
+
+export interface CardCarousel {
+  type: 'carousel';
+  link?: CardLink;
+  media_urls: string[];
+}
+
+export type Card = CardSingle | CardCarousel;
 
 export interface Tweet {
   id: TweetID;
   timestamp: number;
   author: User;
   text: TweetEntity[];
-  annotations?: TweetEntityAnnotation[];
-  referenced_tweets?: ReferencedTweet[];
+  card?: Card;
+  media?: Media[];
 }
